Test case for breakpoint robustness in expressions
This is a certification test for this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/340

"Bug in breakpointing, when an expression combines a ternary and a
table"

The computeBreakpoints function in the expression library performs a
Newton loop to find time-dependent breakpoints in conditional
expressions.

As it performs a Newton loop, it computes an "f" and a "dfdt" derivative
at each iteration.  It was incorrectly attempting to execute the Newton
loop even if the initial "f" and/or "dfdt" were numerically zero.  And,
even if the loop actually made no progress (likely) it would under some
circumstances add a nonsense breakpoint to the breakpoints container.

In most cases, these nonsense breakpoints were discarded and didn't
cause problems.  But in other cases, it could cause the time integrator
to break with an infinite loop in the time integrator.

This netlist is one such test case. In Xyce 7.4, it will go into an
infinite loop and never finish. To mitigate infinite loops (as you'll
get if you attempt to run this in Xyce 7.4), this test
case will uncerimoniously exit at the 200th time step if it hasn't
finished yet.

This commit also updates the tags for other tests of expression
breakpoints, by adding the "expression_breakpoints" tag.