Remove inappropriate "library" tag
The pymi tests all copied the GenExt tags lists, but incorrectly
include "library".

"library" is used by the test suite to signify "this test builds a
test code out of the source tree and links to Xyce libraries".  Any
test that has this tag can't be run except from inside a build
directory, because only build directories retain all the Makefiles and
information needed to locate the source tree.  When we
test *installed* Xyce, we have to exclude those tests.

By including this tag, it would make it impossible to test an
installed Xyce with PyMI should we ever get this capability to the
stage that it could be folded into ordinary regression testing.

I'm removing the tag, even though at the moment it's not interfering
with anything, because we're not testing this stuff in any way.