Create a test for GND synonym in expressions
Issue #320
"Fix handling of ground nodes in expression library to be consistent
with .PREPROCESS REPLACEGROUND"

This is a very simple test that makes sure that when ".PREPROCESS
REPLACEGROUND=true" is present in the netlist, the expression library
will treat nodes names GND, GND!, GROUND as the ground node.

I had previously enabled this, but unconditionally and it wasn't tested
in the regression test suite.  (there were unit tests)

It is now toggled by the PREPROCESS command, which makes it consistent
with the rest of Xyce.