Refine the "warning" tests for this issue
Xyce/Xyce#202
"When multiple instances of the same .param appear in the netlist, Xyce
uses the first value, while Hspice and ngspice use the last value"

This commit updates the test for this issue.  The issue has been nearly
done for a while, but Tom suggested that I improve the "warning" level
tests to use the runAndCheckWarning capability, to ensure that the
proper warning messages were emitted.  This commit accopmlishes that
improvement.