Remove .sens test that is causing xyce to crash
Temporarily disabling a test case that checks for an incomplete .SENS
line. While updating a test case for
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/270, I uncovered an
error case that Xyce isn't handling. The test file AC_doInit3.cir was
updated to check for an error message when an AC sens line defines one
or more objective functions, but does not include a parameter
definitions. Instead of getting an error message, Xyce crashed.
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/371 was created to
address this. This commit comments out this test case until #371 is
resolved.