Adds required:fft tags to several YLIN tests
While configuring a new laptop, I built and test a Xyce development
binary before noticing that FFTW was not installed.  Several regression
test failures occurred due to Xyce not having FFT enabled in its build.
The fft flag was not set during the regression testing, so it should
have filtered out any tests that required FFT.  There were some that
required it, but did not explicitly state the dependency in the tags file.

This fixes that problem for YLIN.