Test of AC .SENS with .PARAM
Xyce#372
"AC sensitivities are incorrect when using .param for sensitivity
parameters"

This test is (like many of them here) based on the simple low-pass
filter circuit.  So, like the others, it is compared against an
analytical solution.

The only difference is the param= field on the .SENS command.  Instead
of:

param=r1:r,c1:c,v1:acmag

this version instead has:

param=rval,cval,vinval

where rval, cval and vinval are .params specified elsewhere in the
netlist, and applied to the r1, c1 and v1 devices, respectively.