Another test of AC .SENS using .param
Xyce#372
"AC sensitivities are incorrect when using .param for sensitivity
parameters"

This is similar to the previous test that I created for this issue.  The
main difference is that this is a nonlinear circuit so it doesn't have
an analytical solution to compare against.  So, instead it is comparing
against a finite difference calculation.