Update the gold standard for Hspice limit test
This is related to Xyce/Xyce#377
"The limit random operator (Hspice 2-argument version) is not correct in
the expression library"

Xyce supports two different versions of limit.  One is the Pspice
version, which takes 3 arguments.  The other is the Hspice version,
which takes 2 arguments.  (technically in Hspice, limit also has an
optional 3rd argument, but I'm ignoring that for now).

This test concerns the 2-argument Hspice version.  That version supports
corner analysis via sampling.  The two arguments are:

limit(nominal,abs_variation)

If running with sampling, then this operator returns either
nominal+abs_variation. or nominal-abs_variation, with equal
probability.

If running without sampling, this operator should return the
nominal value.  However, until today it was returning
nominal+abs_variation.  This test (which is not a sampling test) used
that as the gold standard.

That was incorrect, and the expression library has been updated to
return nominal.  This commit updates the test to expect the same.