Test new ac sensitivity obj func definition rule
Xyce used to have a rule that a .SENS line could not use both acobjfunc
and objvars to define objective functions for doing AC sensitivity
analysis. That restriction was removed as part of the changes for issue
Xyce/Xyce#270. This test checks that the new rule is correctly
implemented (Xyce does not return an error if both objvars and
acobjfunc are used correctly), and that mixing the forms for defining
the objective functions does not change the expected results.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/270