Force these two tests to use the KLU solver
My parameter refacrtoring, which was undertaken for Xyce/Xyce#306,
caused the PCE/NISP2/latch_embedded_nisp_norm.cir to fail for the
"parallel build, serial run" options.

It caused it to fail in part b/c the uncertain parameters are now set up
in a different order.  This is a sensitive circuit, and I have noticed
that the parameter ordering (and thus the matrix ordering)  matters.

The ordering of this problem changed because of my recent parameter
refactoring, which added various sort routines to the function the
sets up uncertain expression-based parameters.  This changed the
order of the blocks in the linear system.

But that order change, even for this sensitive circuit should not have
mattered (much).  But it mattered b/c for the "parallel build, serial
run" tests, it was using AztecOO for linear solves.  That wasn't my
intent, as this test doesn't have a parallel tag.

This circuit does not have a parallel tag but I had forgotten that the
"parallel build, serial run" tests don't look for the parallel tag.
They run the serial tests.

This commit adds the following line to the
PCE/NISP2/latch_embedded_nisp_norm circuit:

.options linsol-es type=klu

And, even though it wasn't failing I have also added this same
line to the PCE/NISP/latch_embedded_nisp_norm circuit.

That test is identical to the one in the NISP2 directory, except for the
manner in which uncertain params are specified.  Although it was not
failing, I hadn't intended for it to use AztecOO either.