Test for the LIMIT sampling operator
Xyce/Xyce#377
"The limit random operator (Hspice 2-argument version) is not correct in
the expression library"

I previousuly set up 2 other tests for this issue.  However, neither of
those tests looked at output statistics.  This test checks that the mean
of limit(mean,+/-delta) is actually equal to the specified mean, when
running a large number of samples.