New test for Xyce/Xyce#306, about slow sampling
This is a test of parameter handling in .sampling.   Prior to recent
fixes for this issue, complicated PDKs with many random expression
operators (agauss, etc) ran really slowly.

This test doesn't look at anything except for runtime.  It is also a
pretty artificial test.  It has a huge number of unused agauss operators
in it, which Xyce 7.4 was unconditionally sampling.

The difference in runtime between today's development branch and
Xyce 7.4 is quite extreme.  In development branch, it runs in 1-2
seconds on my laptop, and in Xyce 7.4, it takes over a minute (~81
seconds on my laptop).

In this test, it fails if the runtime is longer than 10 seconds.