Add a test for issue 396
Issue Xyce#396 was put in yesterday because a published Verilog-A
model for a resistor contained a noise contribution that had an analog
function call on its RHS and Xyce/ADMS generated bad code for it.

This test case demonstrates that the issue is fixed.  It does so
simply by using buildxyceplugin to generate a plugin from that very
model (which is freely downloadable off of github and contains correct
citations in its comments).  Xyce prior to yesterday would fail this
test.

We only test verilog plugins in the "shared_install" builds of Xyce,
but do so nightly.  So this should be sufficient for verification of
Xyce#396.