Test proper code generation for case statements
Xyce#398 was about how Verilog-A case statements are translated to C++
by Xyce/ADMS.  There was a bug that caused bad code because of missing
braces.

These two tests, like those of Xyce#396, simply test that a verilog
plugin will compile from a source code that uses the feature that had
previously been broken, as the failure was a compilation failure
before.

These two tests fail with Xyce 7.4's buildxyceplugin.

One tests use of case in the main analog block, the second use of case
in an analog function.  Both are the same model, just that case
statement has been moved into an analog function for the second.

Xyce#398