Add a test for issue 399
Xyce#399 was about how Xyce/ADMS was gleefully accepting values of
"xyceModelGroup" that don't correspond to actual model groups in Xyce,
and then the C++ code it generated would fail to compile.  The
compilation problem did not give sufficient information for most users
to understand what they did wrong.

The fix was to make Xyce/ADMS exit with an informative error message
rejecting the invalid setting of the module attribute, and telling the
user to remove it.

This test simply makes sure that happens.  The very Verilog-A that a
customer had provided with this problem is passed through
buildxyceplugin, which is expected to fail.  The buildxyceplugin.log
file is then grepped to make sure it contains the correct error
message.

Xyce#399