Start of testing Xyce-Simulink interface
This doesn't work yet but has most of the needed parts.  A shell
script that is callable from the existing test framwork can
run a Simulink problem in Matlab which calls Xyce.  The results
from Xyce are not checked and the command "matlab" must be on
the path (Not sure what I'll do under Windows where it's
probably "matlab.exe").

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/416