Add simulink test tag if matlab is available
If matlab is in the test envrionment then add the simulink based tests
It would be better if this script could also verify that the
xyce-simulink library was built.  That would entail looking for
the file xyce_sfunction.mex* in either the BUILD/utilsSimulinkInteface
directory OR in the INSTALL/lib directory.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/416