Move matlab commands to a separate file
I was having problems with the matlab commands that needed quotes
in them getting the quotes removed by the time the Mathworks script
matlab invoked the Matlab binary MATLAB.  So the matlab commands
now are in their own file and that cleans up how matlab is
invoked from the command line.  Manually running this from the
command line works but from within a perl script it just hangs.
So this still needs more work.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/416