Exclude test until update suggestedXyceTagList
I realized that suggestedXyceTagList will pull in this test if
matlab is present but that doesn't assure that the simulink
library has also been compiled.  This is a situation that the
Charon team may run into.  So for now I'll exclude this
test until I can properly fix the suggestXyceTagList

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/416