Create a new HB test for bug 407
The original test for bug 407 with HB analysis was not set up
correctly. With the parameters for driving source in the circuit,
this linear circuit needs thousands of harmonics in order to generate
correct answer. And the result would be numerically sensitive due
to the sharp transitions which is not well represented in frequency
domain. The circuit did not use enough harmonics and the result was
not correct.

The main requirement for this test is to generate results with
non-zero values at all harmonics. It is better to create a nonlinear
circuit for this purpose instead of using a linear circuit. This commit
creates a new test for bug 407 son. The results are verified with
transient simulation results

Xyce/Xyce#403