Add a new HB output test
The circuit tests different format of HB output for voltage difference.
However, the old circuit did not use enough harmonics and the result
was not correct. In order to create correct result, it needs at least
tens of thousand harmonic. It is not necessary to run a long simulation
for this simple linear circuit to test HB output. In order to have meaninful
phase result, a new nonlinear test is used. The nonlinearity in the circuit
creates non-zero values at higher harmonics. The results are verified by
both the convergence study and by comparing with transient simulation results

Xyce/Xyce#403