Dynamically add to matlab search path for test
The last issue to making this test work on other machines in
general was to dynamically set the matlab search path to where
Xyce's libraries and the needed Simulink definition file are.
Originally I had a fixed file that contained the needed Matlab
code but now I write out a custom file generated by the test
script.  This allows the needed path to be set dynamically
when the test is run and as a bonus the change doesn't
permenently stay set in matlab on the next run.  With that
I'll remove the test from the exclude file so it will be
run if/when we have a nightly build that and use it.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/416