Update gold std for suckt multiplier toggle
Xyce/Xyce#267 - subcircuit multipliers

As part of my work for this issue, I added a new command line option to
allow the user to disable implicit subcircuit multipliers.  This is
mainly for debug purposes.  But, now that this option is present, the
help menu gold standard for this test needed an update.