Add another subckt M= test
Xyce/Xyce#267 - subcircuit multipliers

This test is similar to the first one, except that the subcircuit
contains a diode instead of a resistor, and it is a transient rather
than dc calculation.

Like the previous test, there are 4 equivalent circuits that
should all produce the same answer.  Different aspects of precedence are
tested.