Modify the multi-tone HB test to test the initialization of non transient assisted HB
The multi-tone HB test using AFM based method is modified to test
the efficiency of the initialization of HB methods when transient
analysis is not used for intial guess. In this case, the initialization
should be fast. But before the improvement of the initialization,
it can take the majority of total simulation time. The number of
harmonics used in the simulation is increased to demonstrate the
performance enhancement

Xyce/Xyce#326