Add tests checking that the warning has been fixed.
This test checkes that the warning message from the
mutual inductors is NOT emitted for the test circuit.  When
the device connectivity map was wrong, Xyce would emit
an warning. Now that the device connectivity map is correct
the warning should not be emitted.  The test is a bit
clunky in that the routines in Tools.pm "checkError" look
for a given message in Xyce's output.  But this test needs
to verify that the warming message is no longer there.
Thus there is a lot of error inversion in getting results
back from $Tools->checkError();

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/2