Adds test for "Directory node not found" developer error
The "Directory node not found" developer error has been addressed in
the code to provide a more useful error message, like:

    Device XCSD:G_ABMI1 refers to unknown solution node XCSD:V_V_LFWD

This change also required modifications to the parallel directory
object so that the error behavior is the same between serial and
parallel runs.  This test will check for the new, more useful, error
message in serial and parallel.

Xyce/backlog/project-backlog#151