Add test cases for BSIM4 versioning
We now respect the "version" model parameter of BSIM4 and do our
"best" to use the version requested.

Our "best" at this point is:
  - if the user asks for anything older than 4.6.1, use 4.6.1
  - if the user asks for anything newer than 4.7, use 4.7
  - if the user asks for something in between 4.6.1. and 4.7, use
  4.6.1

This commit adds two test cases.

bsim4_test1 tests that Xyce
emits warnings when a user requests an unsupported version, or uses a
parameter that is not supported in the version requested.

inverter_transient_xyce runs two netlists that differ only by which
version of BSIM4 they're running.  In this particular case, we know
that 4.7 and 4.6.1 do NOT produce matching results, and the test case
makes sure they don't match.  It is a failure if they DO match.

With these tests, we confirm that Xyce is respecting the "version"
model  parameter and actually doing something with it.  And we test
that the "something" is what I said it is.

Xyce/backlog/project-backlog#419