Make new test platform-independent
The XyceRegression/Tools function "runAndCheckWarnings" makes the
assumption that all lines of warnings it is looking for will always
appear in the same order on every platform.  This is not true of
device warnings that come from different netlist lines, and may not be
true in parallel.

A separate function, "runAndCheckGroupedWarnings", does not make the
same assumption and allows you to do things like have multiple groups
of strings to check for --- each string within a group is expected to
appear in the order given, but the groups may appear in any order.
This is the function I should have used in the first place.

In addition to switching to that function, I've added additional
strings to search for (such as the line that says what line of the
netlist contains the problematic .model).

Xyce/backlog/project-backlog#419