Increase output width, stop stripping BUG_
run_xyce_regression and run_xyce_regressionMP have always tried to
format the console output neatly, with the test results always lining
up in a column to the right of the test names.

Very long test names are a problem, though.  So it has always had
logic to truncate the names in some way to fit.  "SandiaTests" is
abbreviated to "ST", and "Certification_Tests" to "C_T".  And until
now, if need be it also stripped out "BUG_" altogether.

This is fine for human readability, but is fatal to machine processing
through summary_dart.pl, the script we use to report to CDash.  This
script needs to be able to find files in the test directory if the
test fails, so it can put useful information into the test result
data.

The "ST" and "C_T" substitutions are easy to deal with ---
summary_dart simply undoes that substitution.  Complete
removal of "BUG_" is not.

I am removing the code that deletes "BUG_" from the "supershortdir"
name.

I am also changing the default "textwidth" from 50 to 65.  This
textwidth was originally to make the entire test summary line fit into
a 100 character wide window.  But since it was chosen, the timing data
has increased in width so much that it no longer fits anyway.
Further, some of us use 80-column windows, and it was always line
wrapping anyway.  With this change, the entire test summary *EXCEPT*
timing fits on an 80-column line, and the entire timing summary wraps
to the next line.

With these two changes, fewer tests will get their test names reduced
in size with text substitutions, NO test will be reduced by having
"BUG_" completely removed from its name, and the lines will still look
reasonably neat on a reasonably sized window.

And most importantly, it will be possible for summary_dart to find
files it needs to report useful information when a test with a longish
name fails.

Unfortunately, the test script still has extraordinarily convoluted
logic to handle *really* long names, and in some cases will have the
audacity to truncate those names to make them fit in the text with
with at least a few dots in between the test name and the result.
This would be impossible to deal with in summary_dart, and a future
hacking run (and a new backlog item) should deal with that.  But for
the current issues, the changes made here should be good enough.

Xyce/backlog/project-backlog#103