Add tests of BSIM4 noise multiplicity
This commit tests both versions of BSIM4 for correct implementation of
multiplicity in the noise models.

The four tests "noise1" through "noise4" for 4.6.1 and 4.7.0 are
patterned after tests from the BSIM group and cover the four
combinations of fnoimod/tnoimod that were accepted by 4.6.1.

But the BSIM group does not provide any good tests of tnoimod=2, the
new model introduced in 4.7.0.  So I had to make one up.  It is a copy
of the gain-stage test Garrick used to test tnoimod=2.  It is not a
completely adequate test, though, as it is missing some of the
parameters that tnoimod=2 uses, and therefore those terms are simply
zero when the code computes noise.

All of these tests follow the same pattern:  a netlist with 6
identical mosfets in parallel is run as a baseline, and then another
netlist with a single mosfet that has "M=6" on its instance line is
run.  The two results are compared and should match.

These tests all fail with code prior to my fix of the BSIM4 devices,
and pass afterwards.

Xyce/backlog/bugs#8