Turn off Auger recombination for this test
I recently discovered that Auger recombination was scaled incorrectly in
the various TCAD/PDE devices.  This was a low consequence mistake as
those devices are not used much, but fixing it was the correct thing to
do.  Due the scaling mistake, the Auger term was effectively zero
before my fix.

So far, this change had not broken any tests, except for this one.  That
suggested that most of the PDE based tests didn't go to high enough
injection for Auger to matter much.

This test case broke on two platforms, suggesting that for
this test, the currents are just high enough to sometimes cause a
comparison failure.

In order to preserve regression, and also to make sure that the "auger"
parameter is exercised in the test suite, I have added "auger=false"
to this netlist.  That will remove Auger recombination from this
calculation, and should make it produce the same answer as before.  That
*should* mean that the gold standard can stay the same without being
updated.

This commit also adds the "pde" tag to this test.

Note, this is a very old test from nearly 20 years ago.  When this
circuit is run in today's development branch, it requires about 1/3 the
time steps as the gold standard.  This means that the gold standard was
probably created with our original (1st out of 3) time integrator. Looking
thru the commit log, I also notice that the tolerance on this test has
been loosened at least once in the past.  So, arguably it is overdue
for a gold standard update anyway.  The main point of this test was
a parsing issue, so the precise gold standard isn't that
important with respect to that original intent.

If this commit doesn't fix the test failures, then I will just update
the gold standard and move on.

 Xyce/backlog/project-backlog#448