Add auger=false, correctly this time.
Yesterday I added auger=false to this test, so that it would (hopefully)
still pass regression testing.  It started failing when I corrected a
scaling mistake in Auger recombination.

However, I noticed today that the test was still failing in nightly
testing.  Then I took a closer look and realized that while I had added
"auger=false" to the netlist, I added it in the wrong place.  Xyce was
printing a warning and running without it, producing the same failing
result as before.

This commit puts "auger=false" in ther correct place.