Disable Auger recombination for this test
I recently fixed a scaling mistake in Auger recombination for the TCAD
devices.    That fix mostly didn't affect the TCAD device tests, as most
of them don't go up to high enough injection levels for Auger to matter.

Howveer, there were two exceptions, and this test was one of them.  As
the scaling mistake was rendering the Auger recombination term so small
as to be essentially zero, the easiest way to fix this test is to turn
off Auger for this calculation.  So, this commit adds "auger=false" to
the TCAD device parameters.

Xyce/backlog/project-backlog#448