Remove build steps with CMake builds of Xyce
All of the tests modified in this commit use small binaries that link to
XyceLib. The source files for the binaries are in subdirectories of
`Xyce/src/test/`. Prior to this change, the tests, using the approach taken
with the Autotools-based build system, built the binaries at test time. That
approach is no longer viable with the CMake system, as the builds sometimes
step on each other. It's kind of bad form to build binaries while testing,
anyway.

Now the tests look to see if CMake was the build system, and don't try to build
anything if that is the case. However, to keep all the test binaries from being
built every time Xyce is built, I put them behind the `BUILD_TESTING` flag.
That is the variable CMake uses for unit tests (in fact, some of these tests
could be changed into unit tests).

If Xyce is not built with `BUILD_TESTING=TRUE`, all these tests will fail.

Xyce/backlog/bugs#3