Modify this test to disable Auger recombination
I recently fixed a scaling mistake in the PDE (TCAD) devices for Auger
recombination.  It broke several tests including this one.  I didn't
notice this one initially because it is a weekly test.

Similar to my other Auger-related test fixes, the solution for this one
was to disable Auger recombination by adding "auger=false" to the two
PDE devices in this file.

For the 2D device (which is the TCAD device used by this test), I had
not set up Auger recombination to be optional.  So, for this to work I
had to also add this parameter to that device.

Xyce/backlog/bugs#9
Xyce/backlog/project-backlog#448