Update BSIM4_Versioning test after 4.8.2 addition
Now that Xyce supports BSIM4.8.2, this test case needed updating. It
was previously looking for a warning that 4.8 was not supported and
that 4.7.0 would be used instead.  Now, 4.8 *IS* supported (it uses
the 4.8.2 code, which contains conditionals to make 4.8 and 4.8.2
different).

The test case now asks for 4.9, which is not supported and falls back
to 4.8.2.

Xyce/backlog/project-backlog#352