Add BSIM4 v4.8.2 benchmarking tests
These 14 tests are the 14 DC tests from the BSIM group's
"Benchmarking" set provided with BSIM 4.8.1.  NOT 4.8.2, which
actually doesn't contain a benchmark suite.

The tests have been modified slightly to output the sweep variables
along with the current, because xyce_verify requires that.  Further,
the DC sweeps have been modified to make Xyce output the same set of
data points that SPICE3F5 does.  SPICE3F5 uses a formulat for
computing DC sweep steps that is prone to roundoff and dropping the
last point requested, and all of the BSIM group's output files have
had that happen.

The "gold" standards are the BSIM group's SPICE3 output files that
have been munged to look like Xyce output (the sweep variable columns
have been added and Xyce-style headers attached, and all SPICE3F5
cruft removed).  The actual column of current output is the BSIM
group's data, the first sweep variable's column is their output, but
the second sweep variable's column just been cut and pasted from
Xyce's output.

All 14 of these tests pass when run with VERSION set to 4.8.1 (as it
was when the BSIM group wrote it), but also pass if VERSION is set to
4.8.2 (as committed here, to silence Xyce warnings about versions) or
even 4.7.0.  They fail horribly if VERSION is set to 4.6.1 in the
model cards, demonstrating that 4.7 and 4.8 are NOT drop in
replacements for 4.6.1 when 4.6.1 data is the "gold" standard, and
4.6.1 will produce wrong results relative to 4.8 or 4.7 gold standards.

Xyce/backlog/project-backlog#352