Send "which" errors to /dev/null
suggestXyceTagList.sh calls "which" on two codes, ADMS and matlab, to
see if they exist in the path and to (possibly) enable or disable
tests based on the results.

In both cases, it redirect's "which" standard output to /dev/null, but
does not redirect its standard error.

On my own primary system, which does not have matlab, this is fine
because apparently "which" sends the error message "no matlab
found..." to stdout.

But on my Linux laptop, it spews that message to standard error, and
therefore blats it on my screen where I don't care to see it.

It no longer does so, and should not do so if a user doesn't have
ADMS, either.