Fix the testBlockLinearSystems test for CMake
This test tried to rebuild the binary under CMake, which causes
problems during testing (see Xyce/backlog/bugs#3). Also, the unit test
version had to be rewritten to have the matrix market files in the Xyce
repository. Since it is bad form to have the files in two places, I
reversed the copy operation so it copies the files from the Xyce repo
into the testing directory.

Xyce/backlog/bugs#14