Test case for issue 451
Xyce/backlog/project-backlog#451
"Allow internal subcircuit nodes to be accessed from netlist top level
(not via interface nodes)"

It turned out that this issue didn't require any code changes.  Xyce
already had this capability.  However, the documentation indicated that
it did not, and also it wasn't  (to the best of my knowledge)
demonstrated in the test suite.

This commit adds a single test that demonstrates this feature.  I have
previously modified the documentation.