Remove lambertw from test netlists
The test "diode_direct.cir" in this directory made use of the diode's
"lambertw" parameter and set it to 1.  lambertw has been removed
from the diode as part of this branch work.

But diode_direct.cir is NOT tested and in fact every file in this
directory is excluded other than diode_forTimurROL.cir, which is
tested nightly in TDEV testing.  So the correct fix for
diode_direct.cir is to remove it altogether.

diode_forTimurROL.cir is run in nightly ROL testing, but its use of
lambertw is merely to set it to zero, which was always the default
anyway and disabled use of lambert-w.  So the right fix for this
netlist is to remove "lambertw=0" from the diode line.

Xyce/backlog/project-backlog#472