Fix non-standard attribute usage
We just discovered another ADMS dinosaur that crept into all of our
Verilog-A examples:  ADMS accepts attribute lists that are space
separated, but the Verilog-A standard says they should be comma
separated.

Many of our toy examples were taking advantage of this non-standard
usage simply because ADMS accepted it without complaint.

Writing portable Verilog-A, however, requires that they be comma
separated, and we don't want to keep examples around of how to do it
non-portably.