Fix non-standard attribute usage
We just discovered another ADMS dinosaur that crept into all of our
Verilog-A examples:  ADMS accepts attribute lists that are space
separated, but the Verilog-A standard says they should be comma
separated.

Many of our toy examples were taking advantage of this non-standard
usage simply because ADMS accepted it without complaint.

Writing portable Verilog-A, however, requires that they be comma
separated, and we don't want to keep examples around of how to do it
non-portably.

I had tried this in a previous commit but botched it because all of
the toys were using a macro to set attributes, and just adding a comma
didn't work.  This does.

Note that the diode clipper gold standard has changed.  That's because
the diode had an error in it in the expression vcrit that was fixed in
the version of the diode in the Xyce/utils/ADMS/examples/toys
directory and only now got copied here.  That changed the output
slightly.

Xyce/backlog/project-backlog#486