Fix inappropriate path hardcoding
The test Certification_Tests/BUG_1325_SON/inc_lib_file_absolute_path
requires that on Windows a Cygwin path be replaced with a
Windows-style path.

Ever since commit e414bba13 in 2020, the test has incorrectly been
assuming that if it's running on Windows under cygwin then the paths
to the files it needs are all on "C:".

That is no longer true, and the Jenkins workspaces are now on E:.

The perl script should not have specifically been looking for
"/cygdrive/c" and replacing with "C:"  but for "/cygdrive/([a-z])" and
replacing with whatever is found in the last position and a colon.

This commit makes the substitution general.

This is still not the "right" way to do this test, but it will work.

Mixing Windows and Cygwin paths is evil, but a necessary evil.