Change hard-coded drive letter
The test
Certification_Tests/BUG_1325_SON/Win/drive_letter_no_slash_path is
ill-conceived as it requires an include file to have a hard-coded,
windows-style path with a drive letter in it.

The test was hard-coded to assume C:.  This is an invalid assumption,
and for some time now we've had our tests on E:.

I don't know exactly why this test has not been showing up as a
failure before now, but it does now.

I am changing the hard-coded drive letter to E: so that this test will
stop failing, but this is a band-aid solution to a femoral bleed.

Since the test is trying to check that Xyce is correctly handling the
case where a drive letter is specified but no path given (which
basically is supposed to mean "the current directory", I guess), what
should probably happen is that the drive letter on which the
file *actually* exists should be substituted at run time into the
include files, and *then* the test run.

But I'm not changing that.  I'm just band-aiding the hard-coded path
and leaving the armed landmine in place.