Revert "Change hard-coded drive letter"
This reverts commit 224c6d66af9e0038e647e0db8e9feb85ec13ff54.

This is getting ridiculous.

I originally changed the hard-coded drive letter of this test from C:
to E:, only to find that it isn't right, either.  The workspace is in
fact on C:.

I don't know why it failed before saying it couldn't find C:<filename>
but it was not right to have changed it to E:.

This test is badly designed and is testing weird windows-specific
include file paths, and they might be broken.  But in the case of the
failures earlier today using E: is wrong.