Exclude a windows test incompatible with our setup
The test
Certification_Tests/BUG_1325_SON/Win/drive_letter_no_slash_path tests
a quirky windows-specific use of relative paths in included files.

The top-level netlist drive_letter_no_slash_path.cir includes a file
sub1/dlns_path.

dlns_path includes two files using .inc and .lib, but uses paths that
have a Windows drive letter and a file name with no slashes in it.
This is supposed to be resolved as a relative path to the current
working directory in that drive, which in this case is supposed to be
the same directory as drive_letter_no_slash_path.cir.

This works just fine when you run Xyce natively, and even works
properly if you run it in Cygwin.  The thing it is intended to test
works.

But as written, the test case is too primitive and too highly
dependent on the way the test machine was configured when the test was
written --- it presumes that the workspace is in C:, and that used to
be true.  As a result, it has begun to fail when we run it through the
regular nightly testing system on our current configuration.

Right now, our test machine has a second disk, E:, that has a
mount-point that makes it *look* like it is a subdirectory of C:.
That is supposed to be completely transparent, but apparently isn't.

Recent work on our test set-up has changed how we invoke
run_xyce_regression --- all Windows testing is done under Cygwin (so
our Unix-style scripts will work), but we used to use a very funky
hybrid of Windows and Unix-style pathing that was fragile.  We now
strictly use  Cygwin paths ("/cygdrive/c/...") consistently.

As a result, something subtle has broken about this test --- while we
are telling our scripts that they should use "/cygdrive/c" paths, in
reality the workspaces physically reside on E:, which has a mount
point under C:.  Somehow, deep in run_xyce_regression, the current
working directory when the test is being run is being unwound and the
mount point forgotten --- the script now thinks it's really under
"/cygrive/e/...".  And so the relative "C:" path in sub1/dlns_path no
longer works.  Worse, even changing that to "E:" in the included file
doesnt' work.

But only when testing under our nightly/weekly test framework.  If one
actually runs this test by hand outside of our perl script framework
it works just fine.

So the bottom line is this test, which has no associated script and
simply relies on a .cir file working out of the box, is too primitive
to work under the Xyce regression framework unless the machine on
which it is running has no fancy configuration.

Making this test robust to changes in environment would require much
more careful work in constructing a proper test script to go along
with it, one that properly unwinds the current directory and
constructs include files that respect the environment they're running
in instead of presuming how it's set up.

For this reason, after spending an hour or so with David trying to
figure out what the heck was going on all of a sudden to make this
fail, I have concluded that this test must be excluded from testing
for the time being as too simplistic and fragile.

If and when we ever return to a configuration where Jenkins workspaces
are no longer on a different drive than everything else and go back to
being on C:, the test can be re-enabled by removing this "exclude"
file.

Whether this happens because we replace the Windows test hardware or
because we move to a containerized Windows testing system is a
question to be resolved later.

This exclusion will at least end one of the many annoying and
meaningless failures we're seeing on Windows these days.